+++ client/app/reader/DocumentSearch.jsx
@@ -5,7 +5,7 @@ import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { getTextSearch, getTextForFile } from './selectors';
 import SearchBar from '../components/SearchBar';
-import { searchText, getDocumentText } from './actions';
+import { searchText, getDocumentText } from './Pdf/PdfActions';
 import _ from 'lodash';
 
 export class DocumentSearch extends React.PureComponent {
