+++ client/app/reader/PdfUI.jsx
@@ -9,7 +9,7 @@ import Pdf from './Pdf';
 import DocumentCategoryIcons from './DocumentCategoryIcons';
 import { connect } from 'react-redux';
 import { selectCurrentPdf, stopPlacingAnnotation, resetJumpToPage,
-  togglePdfSidebar, rotateDocument } from '../reader/actions';
+  togglePdfSidebar, rotateDocument } from './PdfViewer/PdfViewerActions';
 import { docListIsFiltered } from '../reader/selectors';
 import { DownloadIcon, FilterIcon, PageArrowLeft, PageArrowRight, LeftChevron,
   ExternalLink, FitToScreen, Rotate } from '../components/RenderFunctions';
