+++ client/app/reader/PdfUIPageNumInput.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
-import { jumpToPage } from './actions';
+import { jumpToPage } from './PdfViewer/PdfViewerActions';
 import { isValidWholeNumber } from './utils';
 import TextField from '../components/TextField';
