+++ client/app/reader/SideBarComments.jsx
@@ -7,9 +7,8 @@ import { plusIcon } from '../components/RenderFunctions';
 import Button from '../components/Button';
 import _ from 'lodash';
 import { INTERACTION_TYPES } from './analytics';
-
 import { updateNewAnnotationContent, createAnnotation, stopPlacingAnnotation,
-  startPlacingAnnotation } from '../reader/actions';
+  startPlacingAnnotation } from '../reader/PdfViewer/PdfViewerActions';
 
 class SideBarComments extends PureComponent {
   handleAddClick = (event) => {
